Rename EnvNav to ServiceNav and move Integrations link
Renamed `EnvNav` component to `ServiceNav` and updated its imports. Moved the `Integrations` link from `OrgNav` to a newly created `EnvNav` component. Updated navigation to reflect these changes.
